refactor MVVC and create MVVC+CollectionView
