+++ quavi/quavi/Controllers/MapViewController.swift
@@ -13,8 +13,10 @@ import MapboxNavigation
 import MapboxDirections
 
 class MapViewController: UIViewController {
-
+    
+    // MARK: -PROPERTIES
     var sampleData = POI.pointsOfinterest
+    var selectedRoute: Route?
     
     // MARK: - VIEWS
     lazy var mapView = MapView(frame: view.bounds)
@@ -30,10 +32,6 @@ class MapViewController: UIViewController {
         image.isUserInteractionEnabled = true
         return image
     }()
-    
-    
-    // MARK: - Internal Properties
-    var selectedRoute: Route?
     
     // TODO: Make Category enum case iterable and load directly, don't need this property. For MVP we can remove enum and get categories directly from the loaded tours.
     let sampleCategoryData: [CategoryData] = [
@@ -55,9 +53,7 @@ class MapViewController: UIViewController {
     let sliderViewHeight: CGFloat = 900
     
     var currentSelectedCategory: String = Enums.categories.History.rawValue {
-        didSet {
-            poiTableView.reloadData()
-        }
+        didSet { poiTableView.reloadData() }
     }
     
     // MARK: - Lifecycle Functions
@@ -78,7 +74,6 @@ class MapViewController: UIViewController {
         addSliderViewConstraints()
         loadGestures()
         self.startNavigationButton.addTarget(self, action: #selector(startNavigationButtonPressed), for: .touchUpInside)
-
     }
     
     override func viewWillAppear(_ animated: Bool) {
@@ -88,51 +83,6 @@ class MapViewController: UIViewController {
     
     //MARK: -PRIVATE FUNCTIONS
     
-    //MARK: -OBJ-C FUNCTIONS
-    @objc func tvCellSectionButtonPressed(sender: UIButton) {
-        print(sender.tag)
-        if sampleData[sender.tag].isCellExpanded {
-            sampleData[sender.tag].isCellExpanded = false
-        } else {
-            sampleData[sender.tag].isCellExpanded = true
-        }
-        let incides: IndexSet = [sender.tag]
-        poiTableView.reloadSections(incides, with: .fade)
-    }
+    
     
 }
-
-
-
-//TODO: Make file extension of MapViewController named MapViewController+CollectionView
-extension MapViewController: UICollectionViewDataSource, UICollectionViewDelegate, UICollectionViewDelegateFlowLayout {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        return sampleCategoryData.count
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let category = sampleCategoryData[indexPath.row]
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: Enums.cellIdentifiers.categoryCell.rawValue, for: indexPath) as? CategoriesCollectionViewCell else {return UICollectionViewCell()}
-        
-        cell.setUpCells(cell: cell, data: category)
-        
-        return cell
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
-        return CGSize(width: 100, height: 50)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
-        return UIEdgeInsets(top: 10, left: 10, bottom: 10, right: 10)
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, didSelectItemAt indexPath: IndexPath) {
-        handleCollectionViewCellPressed(item: indexPath.item)
-        print(currentSelectedCategory)
-        
-    }
-    
-}
-
-
